revC2: Part ID removed from silkscreen, added logic block legends.
All part IDs and values are now meant to be printed on the Fab layer as
hardcopy manual assembly and population guide. The space on the
silkscreen used for part IDs was partly traded against logical circuit
block documentation. This is just the first pass and meant as a stepping
stone. The goal is to have the logical functionality be as self
descriptive on the silkscreen as possible.